refactor dqn train script
