+++ rllib_integration/helper.py
@@ -6,9 +6,14 @@
 # This work is licensed under the terms of the MIT license.
 # For a copy, see <https://opensource.org/licenses/MIT>.
 
+import collections.abc
+import os
+import shutil
+
 import cv2
 import numpy as np
-import collections.abc
+
+from tensorboard import program
 
 
 def post_process_image(image, normalized=True, grayscale=True):
@@ -30,6 +35,7 @@ def post_process_image(image, normalized=True, grayscale=True):
     else:
         return image.astype(np.uint8)
 
+
 def join_dicts(d, u):
     """
     Recursively updates a dictionary
@@ -42,3 +48,70 @@ def join_dicts(d, u):
         else:
             result[k] = v
     return result
+
+
+def find_latest_checkpoint(directory):
+    """
+    Finds the latest checkpoint, based on how RLLib creates and names them.
+    """
+    start = directory
+    max_checkpoint_int = -1
+    checkpoint_path = ""
+
+    # 1st layer: Check for the different run folders
+    for f in os.listdir(start):
+        if os.path.isdir(start + "/" + f):
+            temp = start + "/" + f
+
+            # 2nd layer: Check all the checkpoint folders
+            for c in os.listdir(temp):
+                if "checkpoint_" in c:
+
+                    # 3rd layer: Get the most recent checkpoint
+                    checkpoint_int = int(''.join([n for n in c
+                                                  if n.isdigit()]))
+                    if checkpoint_int > max_checkpoint_int:
+                        max_checkpoint_int = checkpoint_int
+                        checkpoint_path = temp + "/" + c + "/" + c.replace(
+                            "_", "-")
+
+    if not checkpoint_path:
+        raise FileNotFoundError(
+            "Could not find any checkpoint, make sure that you have selected the correct folder path"
+        )
+
+    return checkpoint_path
+
+
+def get_checkpoint(name, directory, restore=False, overwrite=False):
+    training_directory = os.path.join(directory, name)
+
+    if overwrite and restore:
+        raise RuntimeError(
+            "Both 'overwrite' and 'restore' cannot be True at the same time")
+
+    if overwrite:
+        if os.path.isdir(training_directory):
+            shutil.rmtree(training_directory)
+            print("Removing all contents inside '" + training_directory + "'")
+        return None
+
+
+    if restore:
+        return find_latest_checkpoint(training_directory)
+
+    if os.path.isdir(training_directory) and len(os.listdir(training_directory)) != 0:
+        raise RuntimeError(
+            "The directory where you are trying to train (" +
+            training_directory + ") is not empty. "
+            "To start a new training instance, make sure this folder is either empty, non-existing "
+            "or use the '--overwrite' argument to remove all the contents inside"
+        )
+
+    return None
+
+
+def launch_tensorboard(logdir, host="localhost", port="6006"):
+    tb = program.TensorBoard()
+    tb.configure(argv=[None, "--logdir", logdir, "--host", host, "--port", port])
+    url = tb.launch()
